[FIX] Per100kConfirmed
Change the type of per100kConfirmed key from Int to Float.

+++ src/schema/covid19-schema.ts
@@ -13,7 +13,7 @@ export const covid19Schema = `
       recovered: BasicStructure
       dead: BasicStructure
       vaccination: Vaccination
-      per100kConfirmed: Int
+      per100kConfirmed: Float
     }
 
     type BasicStructure {
